OK, it's probably going to look like a lot, but I just moved some file folder. elements is now circles, because it's for the circle styles. I moved the Recipe API files into src/components/pages/section-nutrition and put the css files in seperate style folders. I deleted the node_modules before upload to take them out of the repo.
