+++ client/src/App.js
@@ -1,5 +1,5 @@
 import LoginForm from './components/Form/LoginForm';
-//import Page from './components/Page';
+import Page from './components/Page';
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import Login from './components/pages/Login';
 import RegisterForm from './components/Form/RegisterForm';
@@ -7,17 +7,22 @@ import { RecipeForm } from './components/Form/RecipeForm';
 
 function App() {
     return (
-        <BrowserRouter>
+        <>
+        <div>
+            <Page/>
+        </div>
+        
+       <BrowserRouter>
             <div className="App">
                 <Routes>
                     <Route path="/" element={<Login />} />
                     <Route path="/register" element={<RegisterForm />} />
                     <Route path="/login" element={<LoginForm />} />
                     <Route path="/recipe" element={<RecipeForm />} />
-
                 </Routes>
             </div>
         </BrowserRouter>
+        </>
     );
 }
