+++ client/src/components/Page.js
@@ -1,15 +1,19 @@
 import React from 'react';
 import Header from './pages/Header';
-// import Home from './pages/Home';
+//import Home from './pages/Home';
 import Footer from './pages/Footer';
 import Login from './pages/Login';
 
+
+
+
 export default function Page() {
+    
     return (
         <div>
             <Header />
             
-                <Login style={{height: "600px"}} />
+                <Login />
                 {/* <Home /> */}
             
             <Footer />
